+++ package.json
@@ -28,13 +28,12 @@
     "@nestjs/platform-express": "^9.0.0",
     "@nestjs/typeorm": "^9.0.1",
     "@types/bcrypt": "^5.0.0",
-<<<<<<< Updated upstream
-=======
     "@types/cookie-parser": "^1.4.3",
     "@types/multer": "^1.4.7",
->>>>>>> Stashed changes
     "bcrypt": "^5.1.0",
     "class-transformer": "^0.3.1",
+    "class-validator": "^0.14.0",
+    "cookie-parser": "^1.4.6",
     "mysql2": "^3.1.2",
     "passport": "^0.6.0",
     "passport-jwt": "^4.0.1",
