--- src/profile/profile.module.ts
@@ -1,13 +1,0 @@
-import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { Profile } from './entities/profile.entity';
-import { ProfileController } from './profile.controller';
-import { ProfileService } from './profile.service';
-
-@Module({
-  imports: [TypeOrmModule.forFeature([Profile])],
-  controllers: [ProfileController],
-  providers: [TypeOrmModule, ProfileService],
-  exports: [ProfileService],
-})
-export class ProfileModule {}
