+++ src/tweet_profile/profile.module.ts
@@ -1,7 +1,13 @@
 import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { Profile } from './entities/profile.entity';
+import { ProfileController } from './profile.controller';
 import { ProfileService } from './profile.service';
 
 @Module({
-  providers: [ProfileService]
+  imports: [TypeOrmModule.forFeature([Profile])],
+  controllers: [ProfileController],
+  providers: [TypeOrmModule, ProfileService],
+  exports: [ProfileService],
 })
 export class ProfileModule {}
