--- src/user/dto/user.dto.ts
@@ -1,6 +1,0 @@
-export class CreateUserDto {
-  user_email: string;
-  user_name: string;
-  user_password: string;
-  user_birthday: string;
-}
