--- src/user/entities/user.entity.ts
@@ -1,19 +1,0 @@
-import { BaseEntity, Column, Entity, PrimaryGeneratedColumn } from 'typeorm';
-
-@Entity()
-export class User extends BaseEntity {
-  @PrimaryGeneratedColumn()
-  id: number;
-
-  @Column()
-  user_name: string;
-
-  @Column({ unique: true })
-  user_email: string;
-
-  @Column()
-  user_password: string;
-
-  @Column()
-  user_birthday: string;
-}
