--- src/user/user.controller.ts
@@ -1,13 +1,0 @@
-import { Body, Controller, Post } from '@nestjs/common';
-import { CreateUserDto } from './dto/user.dto';
-import { UserService } from './user.service';
-
-@Controller('user')
-export class UserController {
-  constructor(private readonly userService: UserService) {}
-
-  @Post()
-  createUser(@Body() userdata: CreateUserDto) {
-    return this.userService.createUser(userdata);
-  }
-}
