--- src/user/user.module.ts
@@ -1,13 +1,0 @@
-import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { UserService } from './user.service';
-import { User } from './entities/user.entity';
-import { UserController } from './user.controller';
-
-@Module({
-  imports: [TypeOrmModule.forFeature([User])],
-  controllers: [UserController],
-  providers: [TypeOrmModule, UserService],
-  exports: [UserService],
-})
-export class UserModule {}
