--- src/user/user.service.ts
@@ -1,24 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { InjectRepository } from '@nestjs/typeorm';
-import { Repository } from 'typeorm';
-import { CreateUserDto } from './dto/user.dto';
-import { User } from './entities/user.entity';
-
-@Injectable()
-export class UserService {
-  constructor(
-    @InjectRepository(User)
-    private userRepository: Repository<User>,
-  ) {}
-
-  async findByEmail(data: string): Promise<User | undefined> {
-    return await this.userRepository.findOne({ where: { user_email: data } });
-  }
-
-  async createUser(data: CreateUserDto) {
-    return await this.userRepository
-      .save(data)
-      .then((res) => res)
-      .catch((e) => console.log(e));
-  }
-}
